+++ stories/misc.stories.js
@@ -1,6 +1,6 @@
 import { storiesOf } from "@storybook/vue";
 import { withMarkdownNotes } from "@storybook/addon-notes";
-import miscDoc from "../docs/storybook/misc.md";
+import miscDoc from "../docs/components/misc.md";
 
 import PulsatingDot from "../components/PulsatingDot.vue";
